Added better handling to work around a bug in how we set the image compression for jp2 files.

+++ plugins/export/Internet_archive.sample.php
@@ -427,6 +427,7 @@ class Internet_archive extends Controller {
 										}
 									}
 								}
+								$preview->setCompressionQuality($quality);
 								$preview->setImageCompressionQuality($quality);
 								echo " creating $new_filebase".".jp2 (Q=$quality)";
 								$preview->setImageDepth(8);
@@ -437,9 +438,13 @@ class Internet_archive extends Controller {
 								$fs = filesize($jp2path.'/'.$new_filebase.'.jp2');
 								while ($fs < 102400) {
 									$tqual++;
+									if ($tqual == 101) {
+										break;
+									}
 									print " $fs is too small (Q=$tqual)";
 									unlink($jp2path.'/'.$new_filebase.'.jp2');
 									$p2 = clone $preview;
+									$p2->setCompressionQuality($tqual);
 									$p2->setImageCompressionQuality($tqual);
 									$p2->writeImage($jp2path.'/'.$new_filebase.'.jp2');
 									$fs = filesize($jp2path.'/'.$new_filebase.'.jp2');
@@ -1556,7 +1561,9 @@ class Internet_archive extends Controller {
 		// Handle the collection(s) that the book might be in
 		$collections = $this->CI->book->get_metadata('collections'); // Returns an array for multiple valies OR a string
 		if (!is_array($collections)) {
-			$collections = array($collections);
+			if ($collections) {
+				$collections = array($collections);
+			}
 		}
 
 		// Combine the "collection" (singular) metadata because, well, people get confused.
